init data asset
